Add files via upload
not for use, ongoing update of the code using 6 permutations and step by step update of the vector for guas

+++ sk_bitfields.h
@@ -258,11 +258,11 @@ public:
 		else if (R <= 0)SetAll_0();
 		else if (R < 64) {
 			bf.u64[0] = 0;
-			bf.u64[0] = (1 << R) - 1;
+			bf.u64[0] = ((uint64_t)1 << R) - 1;
 		}
 		else {
 			bf.u64[0] = 1;
-			bf.u64[0] = (1 << (R - 64)) - 1;
+			bf.u64[0] = ((uint64_t)1 << (R - 64)) - 1;
 		}
 	}
 	inline void Mask(const int theBit) { BF128 w; w.MaskToBit(theBit); *this &= w; }
